+++ src/constants.rs
@@ -1,2 +1,2 @@
 pub const GRID_SIZE: f32 = 1.0;
-pub const CAM_DISTANCE: i32 = 10;
+pub const CAM_DISTANCE: isize = 10;
